requirements.txt
xlrd
-i https://pypi.org/simple
django==2.2.5
gunicorn==19.9.0
pytz==2019.2
sqlparse==0.3.0
django-bootstrap3
django_select2==6.0.3
mysqlclient
django-ckeditor==5.9
djangorestframework
django-cors-headers
pillow